// convex/_generated/api.d.ts
/* eslint-disable */
/**
 * Generated `api` utility.
 *
 * THIS CODE IS AUTOMATICALLY GENERATED.
 *
 * To regenerate, run `npx convex dev`.
 * @module
 */

import type {
  ApiFromModules,
  FilterApi,
  FunctionReference,
} from "convex/server";
import type * as aiAgent from "../aiAgent.js";
import type * as aiAgentInternal from "../aiAgentInternal.js";
import type * as auth from "../auth.js";
import type * as bugs from "../bugs.js";
import type * as http from "../http.js";
import type * as reports from "../reports.js";
import type * as router from "../router.js";
import type * as testCases from "../testCases.js";

/**
 * A utility for referencing Convex functions in your app's API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
declare const fullApi: ApiFromModules<{
  aiAgent: typeof aiAgent;
  aiAgentInternal: typeof aiAgentInternal;
  auth: typeof auth;
  bugs: typeof bugs;
  http: typeof http;
  reports: typeof reports;
  router: typeof router;
  testCases: typeof testCases;
}>;
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;